+++ lib.commonjs/_tests/create-provider.js
@@ -32,6 +32,13 @@ const ProviderCreators = [
     },
     */
     {
+        name: "ChainstackProvider",
+        networks: ["default", "mainnet", "arbitrum", "bnb", "matic"],
+        create: function (network) {
+            return new index_js_1.ChainstackProvider(network);
+        }
+    },
+    {
         name: "EtherscanProvider",
         networks: ethNetworks,
         create: function (network) {
@@ -137,7 +144,7 @@ function getProvider(provider, network) {
 exports.getProvider = getProvider;
 function checkProvider(provider, network) {
     const creator = getCreator(provider);
-    return (creator != null);
+    return (creator != null && creator.networks.indexOf(network) >= 0);
 }
 exports.checkProvider = checkProvider;
 function connect(network) {