+++ lib.esm/providers/provider.js
@@ -1,5 +1,5 @@
 //import { resolveAddress } from "@ethersproject/address";
-import { defineProperties, getBigInt, getNumber, hexlify, resolveProperties, assert, assertArgument, isError, makeError } from "../utils/index.js";
+import { defineProperties, getBigInt, getNumber, hexlify, isBytesLike, resolveProperties, assert, assertArgument, isError, makeError } from "../utils/index.js";
 import { accessListify } from "../transaction/index.js";
 const BN_0 = BigInt(0);
 // -----------------------
@@ -89,7 +89,7 @@ export function copyRequest(req) {
     if (req.data) {
         result.data = hexlify(req.data);
     }
-    const bigIntKeys = "chainId,gasLimit,gasPrice,maxFeePerGas,maxPriorityFeePerGas,value".split(/,/);
+    const bigIntKeys = "chainId,gasLimit,gasPrice,maxFeePerBlobGas,maxFeePerGas,maxPriorityFeePerGas,value".split(/,/);
     for (const key of bigIntKeys) {
         if (!(key in req) || req[key] == null) {
             continue;
@@ -114,6 +114,20 @@ export function copyRequest(req) {
     }
     if ("customData" in req) {
         result.customData = req.customData;
+    }
+    if ("blobVersionedHashes" in req && req.blobVersionedHashes) {
+        result.blobVersionedHashes = req.blobVersionedHashes.slice();
+    }
+    if ("kzg" in req) {
+        result.kzg = req.kzg;
+    }
+    if ("blobs" in req && req.blobs) {
+        result.blobs = req.blobs.map((b) => {
+            if (isBytesLike(b)) {
+                return hexlify(b);
+            }
+            return Object.assign({}, b);
+        });
     }
     return result;
 }